Refacotred netdev to be just a structure
